+++ kipper/core/src/compiler/analysis/analyser/semantic-checker.ts
@@ -8,13 +8,19 @@ import type { KipperProgramContext } from "../../program-ctx";
 import type {
 	CompilableNodeChild,
 	CompilableNodeParent,
-	ScopeNode,
-	Expression,
 	JumpStatement,
 	ReturnStatement,
+	ScopeNode,
 	VariableDeclaration,
 } from "../../ast";
-import { CompoundStatement, FunctionDeclaration, IdentifierPrimaryExpression, IterationStatement } from "../../ast";
+import { LambdaExpression } from "../../ast";
+import {
+	CompoundStatement,
+	Expression,
+	FunctionDeclaration,
+	IdentifierPrimaryExpression,
+	IterationStatement,
+} from "../../ast";
 import { KipperSemanticsAsserter } from "./err-handler";
 import type { LocalScope, Scope } from "../symbol-table";
 import { ScopeDeclaration, ScopeFunctionDeclaration, ScopeVariableDeclaration } from "../symbol-table";
@@ -194,7 +200,7 @@ export class KipperSemanticChecker extends KipperSemanticsAsserter {
 	 * @since 0.10.0
 	 */
 	public validFunctionBody(body: CompilableNodeChild | undefined): void {
-		if (!body || !(body instanceof CompoundStatement)) {
+		if (!body || !(body instanceof CompoundStatement || body instanceof Expression)) {
 			throw this.assertError(new MissingFunctionBodyError());
 		}
 	}
@@ -206,11 +212,14 @@ export class KipperSemanticChecker extends KipperSemanticsAsserter {
 	 * @returns The parent function if found.
 	 * @since 0.10.0
 	 */
-	public getReturnStatementParent(retStatement: ReturnStatement): FunctionDeclaration {
+	public getReturnStatementParent(retStatement: ReturnStatement): FunctionDeclaration | LambdaExpression {
 		// Move up the parent chain and continue as long as there are parents and the current parent is not a function
 		// declaration. This is to ensure a return statement is always used inside a function.
 		let currentParent: CompilableNodeParent | undefined = retStatement.parent;
-		while (!(currentParent instanceof FunctionDeclaration) && currentParent) {
+		while (
+			!(currentParent instanceof FunctionDeclaration || currentParent instanceof LambdaExpression) &&
+			currentParent
+		) {
 			currentParent = currentParent.parent;
 		}