+++ kipper/core/src/compiler/ast/mapping/ast-node-mapper.ts
@@ -51,6 +51,7 @@ import {
 	VoidOrNullOrUndefinedPrimaryExpressionContext,
 	WhileLoopIterationStatementContext,
 	ObjectPropertyContext,
+	LambdaExpressionContext,
 } from "../../lexer-parser";
 import type {
 	ASTDeclarationKind,
@@ -61,7 +62,6 @@ import type {
 	ASTStatementRuleName,
 } from "../common";
 import type { Declaration, Expression, Statement } from "../nodes";
-import { ObjectProperty } from "../nodes";
 import {
 	AdditiveExpression,
 	ArrayPrimaryExpression,
@@ -105,6 +105,8 @@ import {
 	BitwiseXorExpression,
 	BitwiseShiftExpression,
 	BitwiseAndExpression,
+	ObjectProperty,
+	LambdaExpression,
 } from "../nodes";
 
 /**
@@ -162,7 +164,8 @@ export class ASTNodeMapper {
 		[ParseRuleKindMapping.RULE_bitwiseAndExpression]: BitwiseAndExpression,
 		[ParseRuleKindMapping.RULE_bitwiseXorExpression]: BitwiseXorExpression,
 		[ParseRuleKindMapping.RULE_bitwiseShiftExpression]: BitwiseShiftExpression,
-	} satisfies Record<ASTExpressionKind, typeof Expression<any, any>>;
+		[ParseRuleKindMapping.RULE_lambdaExpression]: LambdaExpression,
+	} satisfies Record<ASTExpressionKind, typeof Expression<any, any, any>>;
 
 	/**
 	 * A mapping matching all {@link ASTStatementKind statement kinds} to their respective constructable AST node
@@ -228,6 +231,7 @@ export class ASTNodeMapper {
 		[ParseRuleKindMapping.RULE_bitwiseAndExpression]: BitwiseAndExpressionContext,
 		[ParseRuleKindMapping.RULE_bitwiseXorExpression]: BitwiseXorExpressionContext,
 		[ParseRuleKindMapping.RULE_bitwiseShiftExpression]: BitwiseShiftExpressionContext,
+		[ParseRuleKindMapping.RULE_lambdaExpression]: LambdaExpressionContext,
 		[ParseRuleKindMapping.RULE_memberAccessExpression]: [
 			// Due to the nature of the parser not handling the notations as one rule, it's an array
 			DotNotationMemberAccessExpressionContext,
@@ -301,7 +305,8 @@ export class ASTNodeMapper {
 		RULE_bitwiseAndExpression: BitwiseAndExpression,
 		RULE_bitwiseXorExpression: BitwiseXorExpression,
 		RULE_bitwiseShiftExpression: BitwiseShiftExpression,
-	} satisfies Record<ASTExpressionRuleName, typeof Expression<any, any>>;
+		RULE_lambdaExpression: LambdaExpression,
+	} satisfies Record<ASTExpressionRuleName, typeof Expression<any, any, any>>;
 
 	/**
 	 * A mapping matching all {@link ASTStatementRuleName statement rule names} to their respective constructable AST