+++ kipper/core/src/compiler/ast/nodes/expressions/function-call-expression/function-call-expression-semantics.ts
@@ -3,7 +3,6 @@
  * @since 0.5.0
  */
 import type { Reference } from "../../../../analysis";
-import type { KipperReferenceable } from "../../../../const";
 import type { Expression } from "../expression";
 import type { ExpressionSemantics } from "../expression-semantics";
 
@@ -16,12 +15,12 @@ export interface FunctionCallExpressionSemantics extends ExpressionSemantics {
 	 * The identifier of the function that is called.
 	 * @since 0.5.0
 	 */
-	identifier: string;
+	identifier?: string;
 	/**
 	 * The function that is called by this expression.
 	 * @since 0.5.0
 	 */
-	callTarget: Reference<KipperReferenceable>;
+	target: Expression | Reference;
 	/**
 	 * The arguments that were passed to this function.
 	 * @since 0.6.0